Add prettier, linter, clean up contracts a little, add some celo registry helpers

+++ contracts/KAsset.sol
@@ -4,22 +4,22 @@ pragma solidity ^0.8.4;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract KAsset is ERC20 {
-	address public kresko;
+  address public kresko;
 
-	modifier onlyKresko() {
-		require(msg.sender == kresko, "Sender not kresko");
-		_;
-	}
+  modifier onlyKresko() {
+    require(msg.sender == kresko, "Sender not kresko");
+    _;
+  }
 
-	constructor(
-		address kresko_,
-		string memory name_,
-		string memory symbol_
-	) ERC20(name_, symbol_) {
-		kresko = kresko_;
-	}
+  constructor(
+    address kresko_,
+    string memory name_,
+    string memory symbol_
+  ) ERC20(name_, symbol_) {
+    kresko = kresko_;
+  }
 
-	function mint(address to, uint amount) external onlyKresko {
-		_mint(to, amount);
-	}
+  function mint(address to, uint256 amount) external onlyKresko {
+    _mint(to, amount);
+  }
 }
