+++ contracts/interfaces/IOracle.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.4;
 
 interface IOracle {
-	function value() external returns (uint);
+  function value() external returns (uint256);
 }
